requirements.txt
opencv-python
pillow
numpy
rembg
moviepy
ffmpeg